requirements.txt
asgiref==3.8.1
attrs==24.2.0
boto3==1.35.79
botocore==1.35.79
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==43.0.1
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.1
django-allauth==65.0.2
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
filelock==3.16.1
git-filter-repo==2.45.0
identify==2.6.3
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
nodeenv==1.9.1
oauthlib==3.2.2
pillow==10.4.0
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.28.0
